update folder not extist error when delete

+++ modules/admin/models/News.php
@@ -136,21 +136,23 @@ class News extends \app\models\News
      * @return mixed
      */
     public static function deleteDir($dirPath) {
-        if (! is_dir($dirPath)) {
-            throw new \InvalidArgumentException("$dirPath must be a directory");
-        }
-        if (substr($dirPath, strlen($dirPath) - 1, 1) != '/') {
-            $dirPath .= '/';
-        }
-        $files = glob($dirPath . '*', GLOB_MARK);
-        foreach ($files as $file) {
-            if (is_dir($file)) {
-                self::deleteDir($file);
-            } else {
-                unlink($file);
+        if(file_exists($dirPath)){
+            if (! is_dir($dirPath)) {
+                throw new \InvalidArgumentException("$dirPath must be a directory");
             }
+            if (substr($dirPath, strlen($dirPath) - 1, 1) != '/') {
+                $dirPath .= '/';
+            }
+            $files = glob($dirPath . '*', GLOB_MARK);
+            foreach ($files as $file) {
+                if (is_dir($file)) {
+                    self::deleteDir($file);
+                } else {
+                    unlink($file);
+                }
+            }
+            rmdir($dirPath);
         }
-        rmdir($dirPath);
     }
     
     /*
